+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_assets.py
@@ -43,7 +43,7 @@ from dagster._core.instance import DynamicPartitionsStore
 from dagster._core.remote_representation.code_location import CodeLocation
 from dagster._core.remote_representation.external import ExternalRepository
 from dagster._core.remote_representation.external_data import ExternalAssetNode
-from dagster._core.storage.batch_asset_record_loader import AssetSummaryRecordLoader
+from dagster._core.storage.asset_summary_record_loader import AssetSummaryRecordLoader
 from dagster._core.storage.event_log.sql_event_log import get_max_event_records_limit
 from dagster._core.storage.partition_status_cache import (
     build_failed_and_in_progress_partition_subset,
