+++ python_modules/dagster-graphql/dagster_graphql_tests/graphql/test_assets.py
@@ -1370,7 +1370,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
         # Test that when partition a is materialized that the materialized partitions are a
         _create_partitioned_run(graphql_context, "partition_materialization_job", partition_key="a")
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         selector = infer_job_selector(graphql_context, "partition_materialization_job")
         result = execute_dagster_graphql(
@@ -1392,7 +1392,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
         # Test that when partition c is materialized that the materialized partitions are a, c
         _create_partitioned_run(graphql_context, "partition_materialization_job", partition_key="c")
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -1436,7 +1436,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             tags={"fail": "true"},
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -1459,7 +1459,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             tags={"fail": "true"},
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -1492,7 +1492,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
         assert not result.errors
         assert result.data
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         stats_result = execute_dagster_graphql(
             graphql_context,
@@ -1606,7 +1606,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             graphql_context, "time_partitioned_assets_job", partition_key=time_0
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         selector = infer_job_selector(graphql_context, "time_partitioned_assets_job")
         result = execute_dagster_graphql(
@@ -1632,7 +1632,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             graphql_context, "time_partitioned_assets_job", partition_key=time_2
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -1660,7 +1660,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             graphql_context, "time_partitioned_assets_job", partition_key=time_1
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -1783,7 +1783,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
         # Test with 1 run on all assets
         first_run_id = _create_run(graphql_context, "failure_assets_job")
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -1816,7 +1816,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             asset_selection=[{"path": ["asset_3"]}],
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -2166,7 +2166,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
                 asset_selection=[AssetKey("multipartitions_1")],
             )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -2286,7 +2286,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
                 tags={"fail": "true"},
             )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -2335,7 +2335,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
                 tags={"fail": "true"},
             )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -2367,7 +2367,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
                 MultiPartitionKey({"date": partition_field[0], "ab": partition_field[1]}),
             )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         result = execute_dagster_graphql(
             graphql_context,
@@ -2408,7 +2408,7 @@ class TestAssetAwareEventLog(ExecutingGraphQLContextTestMatrix):
             MultiPartitionKey({"dynamic": "1", "static": "a"}),
         )
 
-        graphql_context.asset_record_loader.clear_cache()
+        graphql_context.asset_summary_record_loader.clear_cache()
 
         counter = Counter()
         traced_counter.set(counter)
